+++ lib/drag_and_drop_list_expansion.dart
@@ -134,23 +134,7 @@ class DragAndDropListExpansion implements DragAndDropListExpansionInterface {
       for (int i = 0; i < children.length; i++) {
         contents.add(DragAndDropItemWrapper(
           child: children[i],
-          onPointerDown: params.onPointerDown,
-          onPointerUp: params.onPointerUp,
-          onPointerMove: params.onPointerMove,
-          onItemReordered: params.onItemReordered,
-          itemOnWillAccept: params.itemOnWillAccept,
-          sizeAnimationDuration: params.itemSizeAnimationDuration,
-          ghostOpacity: params.itemGhostOpacity,
-          ghost: params.itemGhost,
-          dragOnLongPress: params.dragOnLongPress,
-          draggingWidth: params.itemDraggingWidth,
-          axis: params.axis,
-          verticalAlignment: params.verticalAlignment,
-          dragHandle: params.dragHandle,
-          dragHandleOnLeft: params.dragHandleOnLeft,
-          decorationWhileDragging: params.itemDecorationWhileDragging,
-          dragHandleVerticalAlignment: params.itemDragHandleVerticalAlignment,
-          constrainDraggingAxis: params.constrainDraggingAxis,
+          parameters: params,
         ));
         if (params.itemDivider != null && i < children.length - 1) {
           contents.add(params.itemDivider);